Update config.default.sh
The global API key and email address should stay the same for most people as multiple domains are manged by a single account, the only variable that changes is the Zone ID key for each domain. If a domain is on a different account a user can set the global api key and email for that particular domain to be different, overriding the previously set variable.

I believe this example config makes more sense and is cleaner. I was confused and thrown off by yours and manage several sub-domains for each domain.. I used this format for my config and it worked excellent.