Add support for using API tokens instead of email+global API key
Cloudflare added support for a new type of token which can be scoped
and restricted in usage. For example, it can be set to only allow
editing DNS records in specific zones and from certain IP ranges,
whereas global API keys have full account access.

Documentation on API tokens can be found here:
https://developers.cloudflare.com/api/tokens

The old global API keys are still supported, and will be attempted to
be used if the API token is not configured.